refactor: разобьет код на более мелкие части
